comfy/ai-agent-extension/requirements.txt
aiohttp~=3.9.5
requests~=2.32.3
filelock~=3.15.4
openai-whisper==v20231117
torchaudio
ffmpeg-python==0.2.0